LogMsg Model
Almost finished LogMsg database model
