--- src/server/sql/logbase/get_all_logs.sql
@@ -1,4 +1,0 @@
-
-
--- gets all logs in the database
-SELECT * FROM logbase;
