--- src/server/sql/logbase/get_logs_from_type.sql
@@ -1,9 +1,0 @@
-
-
--- Gets logs in table by date range. This is then ordered by time ascending.
-SELECT 
-  -- Short column names for smaller data.
-  log_type as p, log_message as m, log_time as i
-FROM logbase 
-WHERE log_type = ${logType}
-ORDER BY log_time ASC;
