--- src/server/sql/logbase/insert_new_log.sql
@@ -1,5 +1,0 @@
-
-
---Inserts a new log into the table
-INSERT INTO logbase (log_type, log_message, log_time)
-VALUES (${logType}, ${logMessage}, ${logTime});
