+++ src/server/sql/logmsg/get_all_logs.sql
@@ -1,0 +1,6 @@
+/* This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
+
+-- gets all logs in the database
+SELECT * FROM logmsg;
