+++ src/server/sql/logmsg/get_logs_from_dates.sql
@@ -2,9 +2,9 @@
 
 -- Gets logs in table by date range. This is then ordered by time ascending.
 SELECT 
-  -- Short column names for smaller data.
-  log_type as p, log_message as m, log_time as i
-FROM logbase 
+	-- Short column names for smaller data.
+	log_type, log_message as log_msg, log_time
+FROM logmsg 
 WHERE log_time >= COALESCE(${startDate}, '-infinity'::TIMESTAMP)
-  AND log_time <= COALESCE(${endDate}, 'infinity'::TIMESTAMP)
+	AND log_time <= COALESCE(${endDate}, 'infinity'::TIMESTAMP)
 ORDER BY log_time ASC;
