+++ src/server/sql/logmsg/insert_new_log.sql
@@ -1,0 +1,7 @@
+/* This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
+
+--Inserts a new log into the table
+INSERT INTO logmsg (log_type, log_message, log_time)
+VALUES (${logType}, ${logMessage}, ${logTime});
